requirements.txt
# Multi-Agent PostgreSQL Analysis System Dependencies

# Core LangGraph and orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# LLM Integration
litellm>=1.40.0

# Database connectivity
psycopg2-binary>=2.9.7
asyncpg>=0.28.0

# Async support
aiohttp>=3.9.0
websockets>=11.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# JSON handling
orjson>=3.9.0

# HTTP client
httpx>=0.25.0

# Process management
psutil>=5.9.0

# CLI and logging
click>=8.1.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Web Interface
streamlit>=1.28.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4

# MCP Protocol and Advanced Features
mcp>=1.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
websockets>=11.0.0
pydantic-settings>=2.0.0

# Database Connection Pooling
sqlalchemy>=2.0.0
asyncpg-pool>=0.1.0
psycopg2-pool>=1.0.0

# Monitoring and Metrics
prometheus-client>=0.19.0
prometheus-async>=22.2.0
opencensus-ext-prometheus>=0.8.0

# Enhanced Async Support
asyncio-pool>=0.6.0
aiofiles>=23.2.0

# Security and Rate Limiting
slowapi>=0.1.9
python-jose>=3.3.0
passlib>=1.7.4

# Caching
redis>=5.0.0
aioredis>=2.0.0